requirements.txt
jupyterlab
scikit-learn>=0.23.2
seaborn>=0.11.0
opentsne>=0.4.0
SimpleITK==1.2.4
scikit-image==0.16.2
scipy==1.4.1
pydicom==2.0.0
albumentations==0.4.6
tensorflow-addons==0.11.2
tqdm
gdown==4.2.0
keras

